+++ jablib/src/main/java/org/jabref/logic/importer/fileformat/MedlineImporter.java
@@ -160,6 +160,7 @@ public class MedlineImporter extends Importer implements Parser {
             }
         }
 
+        MedlineDoiCleanup.cleanup(fields);
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         entry.setField(fields);
 
@@ -377,6 +378,7 @@ public class MedlineImporter extends Importer implements Parser {
             }
         }
 
+        MedlineDoiCleanup.cleanup(fields);
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         entry.setField(fields);